Fix type casting for CONTENTS_NODE checks in Vulkan renderer
- Updated multiple files in the Vulkan renderer to explicitly cast CONTENTS_NODE to (int) during comparisons. This change ensures type consistency and prevents potential issues related to type mismatches in the rendering logic.

These modifications enhance the robustness of the Vulkan renderer's node handling.

+++ src/renderers/vulkanrenderer/tr_bsp.c
@@ -1938,7 +1938,7 @@ R_SetParent
 static void R_SetParent( mnode_t *node, mnode_t *parent )
 {
 	node->parent = parent;
-	if ( node->contents != CONTENTS_NODE )
+        if ( node->contents != (int)CONTENTS_NODE )
 		return;
 	R_SetParent( node->children[0], node );
 	R_SetParent( node->children[1], node );
